modals for createlistings and avatar image working, deleted console.logs
