--- js/shared/messages.mjs
@@ -1,1 +1,0 @@
-// error and success messages here? (separate files or together in one file?)
