+++ js/utils/setupChangePhotoModal.mjs
@@ -1,58 +1,41 @@
 import { updateAvatar } from "../components/editProfile/editAvatarPhoto.mjs";
 
-export function setupCreateListingModal() {
-  document.addEventListener("DOMContentLoaded", () => {
-    const openModalBtn = document.getElementById("open-modal");
-    const closeModalBtn = document.getElementById("close-modal");
-    const modal = document.getElementById("createlisting-modal");
-
-    if (openModalBtn && closeModalBtn && modal) {
-      openModalBtn.addEventListener("click", () => {
-        modal.classList.remove("hidden");
-        modal.classList.add("flex");
-      });
-
-      closeModalBtn.addEventListener("click", () => {
-        modal.classList.remove("flex");
-        modal.classList.add("hidden");
-      });
-
-      modal.addEventListener("click", (e) => {
-        if (e.target === modal) {
-          modal.classList.add("hidden");
-        }
-      });
-    }
-  });
-}
-
 export function setupChangePhotoModal() {
   const openBtn = document.getElementById("change-photo-modal-btn");
   const closeBtn = document.getElementById("close-photo-modal");
   const modal = document.getElementById("change-photo-modal");
   const changePhotoForm = document.getElementById("change-photo-form");
   const avatarImg = document.getElementById("profile-avatar");
-  const avatarInput = document.getElementById("avatar-url"); // optional hidden input
+  const avatarInput = document.getElementById("avatar-url"); // hidden input in edit form
+  const messageDiv = document.getElementById("message"); 
 
   if (!openBtn || !closeBtn || !modal || !changePhotoForm || !avatarImg) return;
 
+  // Open modal
   openBtn.addEventListener("click", () => {
     modal.classList.remove("hidden");
     modal.classList.add("flex");
+    changePhotoForm.reset();
+    if (messageDiv) messageDiv.textContent = "";
   });
 
+  // Close modal
   closeBtn.addEventListener("click", () => {
     modal.classList.remove("flex");
     modal.classList.add("hidden");
+    if (messageDiv) messageDiv.textContent = "";
   });
 
+  // click outside modal to close modal
   modal.addEventListener("click", (e) => {
     if (e.target === modal) {
       modal.classList.remove("flex");
       modal.classList.add("hidden");
+      if (messageDiv) messageDiv.textContent = "";
     }
   });
 
+  // Handle avatar update
   changePhotoForm.addEventListener("submit", async (e) => {
     e.preventDefault();
     const urlInput = document.getElementById("new-photo-url");
@@ -71,11 +54,16 @@ export function setupChangePhotoModal() {
       try {
         await updateAvatar(newUrl);
         avatarImg.src = newUrl;
-        if (avatarInput) {
-          avatarInput.value = newUrl;
+        if (avatarInput) avatarInput.value = newUrl;
+        if (messageDiv) {
+          messageDiv.textContent = "Avatar updated successfully!";
         }
-        modal.classList.add("hidden");
-        modal.classList.remove("flex");
+        changePhotoForm.reset();
+        setTimeout(() => {
+          modal.classList.remove("flex");
+          modal.classList.add("hidden");
+          if (messageDiv) messageDiv.textContent = "";
+        }, 1200);
       } catch (error) {
         alert("Could not update avatar: " + error.message);
       }
