+++ pages/contact.html
@@ -74,7 +74,6 @@
           <form novalidate id="contactForm">
             <fieldset>
               <legend class="mb-4 text-center">Send us a message</legend>
-              <!-- id="message" is for displaying messages to the user-->
               <div id="message"></div>
               <!-- Title field-->
               <div class="space-y-2">
