+++ pages/listings.html
@@ -78,7 +78,7 @@
           <!-- Button to open modal -->
           <div class="flex justify-end w-full my-4 pr-4">
             <button
-              id="open-modal"
+              id="create-listing-modal-listings"
               type="button"
               class="px-4 py-1 bg-forest-green hover:bg-light-forest-green rounded-3xl text-earthy-beige text-bold shadow"
             >
@@ -89,7 +89,7 @@
           <!-- Modal Overlay -->
           <div
             id="createlisting-modal"
-            class="fixed inset-0 bg-light-forest-green bg-opacity-50 items-center justify-center z-50 hidden"
+            class="fixed inset-0 bg-light-forest-green/90 items-center justify-center z-50 hidden"
           >
             <!-- Modal Content -->
             <section
@@ -97,7 +97,7 @@
             >
               <!-- Close Button -->
               <button
-                id="close-modal"
+                id="close-createlisting-modal"
                 class="absolute top-4 right-4 text-2xl font-bold"
                 aria-label="Close modal"
               >
